[fix] swap commands
